Start of Hiring and firing
Moved up to "How to interview"
This area seems a bit tricky, as some of this is the 'post' interview
evaluation and/or fits into culture.
I've moved one to culture, since it
deals more with internally growing your team, and only moderately
related to hiring in and of itself.