Presentation from past CTO Summit Sydney
Adding my presentation from a past CTO Summit Sydney on ways to think about Dev Onboarding.  There are 10 questions that help ensure Onboarding program is focused on best outcomes.  It is not a how-to article, as I believe that each org is different and needs to customise the program for its own unique situation.